requirements.txt
# Web框架
fastapi==0.103.1
uvicorn==0.23.2
starlette==0.27.0
pydantic==2.3.0
python-multipart==0.0.6
email-validator==2.0.0

# 身份验证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 数据库
motor==3.3.1
pymongo==4.5.0
redis==5.0.0
beanie==1.21.0

# 缓存与会话
itsdangerous==2.1.2

# 微信API集成
wechatpy==2.0.0
cryptography==41.0.3
requests==2.31.0

# 异步任务处理
celery==5.3.4
flower==2.0.1

# 图像处理
Pillow==10.0.1

# 工具与辅助库
pydantic-settings==2.0.3
tenacity==8.2.3
httpx==0.25.0
ujson==5.8.0
python-slugify==8.0.1
gunicorn==21.2.0

# CORS支持
CORSMiddleware==0.1.2

# 性能优化
orjson==3.9.7
uvloop==0.17.0 ; sys_platform != 'win32' and implementation_name == 'cpython'
httptools==0.6.0

# 日志与监控
loguru==0.7.2

# 测试工具
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
asgi-lifespan==2.1.0

# 代码质量工具
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1 